Fix first source import modal opening early
